feat(hotel-booking-frontend): signup provider created and divided auth provider
