+++ apps/1CD/hotel-booking/frontend/src/app/(client)/layout.tsx
@@ -1,12 +1,18 @@
 import { PropsWithChildren } from 'react';
-import { AuthProvider } from '../../components/providers';
+import { AuthProvider, SignupProvider } from '../../components/providers';
+import Footer from '@/components/Footer';
+import HeaderCheckout from '@/components/HeaderCheckout';
 
 const ClientLayout = ({ children }: PropsWithChildren) => {
   return (
     <AuthProvider>
-      <div className="flex flex-col min-h-screen">
-        <div className="flex-1">{children}</div>
-      </div>
+      <SignupProvider>
+        <div className="flex flex-col min-h-screen">
+          <HeaderCheckout />
+          <div className="flex-1">{children}</div>
+          <Footer />
+        </div>
+      </SignupProvider>
     </AuthProvider>
   );
 };
