+++ apps/1CD/hotel-booking/frontend/src/app/(client)/profile/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Page = () => {
+  return <div className="container m-auto">Page</div>;
+};
+
+export default Page;
