+++ apps/1CD/hotel-booking/frontend/src/components/providers/index.ts
@@ -1,3 +1,4 @@
 export * from './ApolloWrapper';
 export * from './Sidebar';
 export * from './AuthProvider';
+export * from './SignupProvider';
